functions/articles-api/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-cloud-logging==3.*
flask==2.*